requirements.txt

streamlit
joblib
pandas
scikit-learn
pillow
